Follow .NET conventions for test class names
